+++ AngularClient/src/main/core/project.service.ts
@@ -12,11 +12,9 @@ import { ElementField, ElementFieldDataType } from "./entities/element-field";
 import { ElementItem } from "./entities/element-item";
 import { Project } from "./entities/project";
 import { UserElementCell } from "./entities/user-element-cell";
-import { UserElementField } from "./entities/user-element-field";
 import { AppEntityManager } from "./app-entity-manager.service";
 import { AuthService } from "./auth.service";
 import { getUniqueValue } from "../shared/utils";
-import { User } from "./entities/user";
 
 @Injectable()
 export class ProjectService {
@@ -35,55 +33,34 @@ export class ProjectService {
     this.appHttpClient = httpClient as AppHttpClient;
   }
 
-  createElement(initialValues: Object) {
+  createElement(initialValues: {}) {
     return this.appEntityManager.createEntity("Element", initialValues);
   }
 
-  createElementCell(initialValues: Object) {
-
-    const elementCell = this.appEntityManager.createEntity("ElementCell", initialValues) as ElementCell;
-
-    // If DataType is decimal, also create "User element cell"
-    if (elementCell.ElementField.DataType === ElementFieldDataType.Decimal) {
-
-      elementCell.DecimalValueTotal = 0; // Computed field
-      elementCell.DecimalValueCount = 1; // Computed field
-
-      const userElementCellInitial = {
-        User: this.authService.currentUser,
-        ElementCell: elementCell,
-        DecimalValue: 0,
-      } as any;
-
-      this.appEntityManager.createEntity("UserElementCell", userElementCellInitial) as UserElementCell;
-    }
-
-    return elementCell;
+  createElementCell(initialValues: {}) {
+    return this.appEntityManager.createEntity("ElementCell", initialValues) as ElementCell;
   }
 
-  createElementField(initialValues: Object, rating: number = 50) {
+  createUserElementCell(elementCell: ElementCell, decimalValue: number | null) {
 
-    const elementField = this.appEntityManager.createEntity("ElementField", initialValues) as ElementField;
-
-    // If RatingEnabled, also create "User element field"
-    if (elementField.RatingEnabled) {
-
-      elementField.RatingTotal = 0; // Computed field
-      elementField.RatingCount = 1; // Computed field
-
-      const userElementFieldInitial = {
-        User: this.authService.currentUser,
-        ElementField: elementField,
-        Rating: rating
-      };
-
-      this.appEntityManager.createEntity("UserElementField", userElementFieldInitial) as UserElementField;
+    if (elementCell.ElementField.DataType !== ElementFieldDataType.Decimal) {
+      throw new Error(`Invalid field type: ${elementCell.ElementField.DataType}`);
     }
 
-    return elementField;
+    const initialValues = {
+      User: this.authService.currentUser,
+      ElementCell: elementCell,
+      DecimalValue: decimalValue
+    };
+
+    return this.appEntityManager.createEntity("UserElementCell", initialValues) as UserElementCell;
   }
 
-  createElementItem(initialValues: Object): ElementItem {
+  createElementField(initialValues: {}) {
+    return this.appEntityManager.createEntity("ElementField", initialValues) as ElementField;
+  }
+
+  createElementItem(initialValues: {}): ElementItem {
     return this.appEntityManager.createEntity("ElementItem", initialValues) as ElementItem;
   }
 
@@ -100,212 +77,82 @@ export class ProjectService {
     return project;
   }
 
-  createDefaultProject(currentUser: User, projectName: string, mainElementName: string, addImportanceIndex: boolean, numberOfItem: number ): Project {
+  createProjectParentChild() {
 
     // Project
-    const project = this.appEntityManager.createEntity("Project", {
-      User: currentUser,
-      Name: projectName,
-      Origin: "http://"
-    }) as Project;
-
-    // Main element
-    const element = this.createElement({
-      Project: project,
-      Name: mainElementName
-    }) as Element;
-
-    // Importance field
-    var importanceField: ElementField = null;
-
-    if (addImportanceIndex) {
-      importanceField = this.createElementField({
-        Element: element,
-        Name: "Importance Field",
-        DataType: ElementFieldDataType.Decimal,
-        UseFixedValue: false,
-        RatingEnabled: true,
-        SortOrder: 1
-      }) as ElementField;
-    }
-
-    // Items, cells
-    for (var i = 0; i < numberOfItem; i++) {
-
-      const item = this.createElementItem({
-        Element: element,
-        Name: `Item ${ i + 1 }`
-      }) as ElementItem;
-
-      if (addImportanceIndex) {
-        this.createElementCell({
-          ElementField: importanceField,
-          ElementItem: item
-        });
-      }
-    }
-
-    // Return
-    return project;
-  }
-
-  createTwoElemenstSample() {
-
-    const numberOfItems = 2;
-
     const project = this.createDefaultProject(
-      this.authService.currentUser, // User
-      "Two Elements Sample", // Project Name
-      "Organization", // Main Element Name
-      false, // Add Importance Item
-      numberOfItems
+      "Parent - Child", // Project name
+      "Parent", // Main element name
+      2
     );
 
-    // Lisance element
-    const lisanceElement = this.createElement({
+    // Child element
+    const childElement = this.createElement({
       Project: project,
-      Name: "License"
+      Name: "Child"
     }) as Element;
 
-    // Fields
-    const rightToUseField = this.createElementField({
-      Element: lisanceElement,
-      Name: "Right to Use",
-      DataType: ElementFieldDataType.String,
-      UseFixedValue: false,
-      RatingEnabled: false
-    }) as ElementField;
-
-    const rightToCopyField = this.createElementField({
-      Element: lisanceElement,
-      Name: "Right to Copy",
-      DataType: ElementFieldDataType.String,
-      UseFixedValue: false,
-      RatingEnabled: false
-    }) as ElementField;
-
-    const rightToModifyField = this.createElementField({
-      Element: lisanceElement,
-      Name: "Right to Modify",
-      DataType: ElementFieldDataType.String,
-      UseFixedValue: false,
-      RatingEnabled: false
-    }) as ElementField;
-
-    const rightToSellField = this.createElementField({
-      Element: lisanceElement,
-      Name: "Right to Sell",
-      DataType: ElementFieldDataType.String,
-      UseFixedValue: false,
-      RatingEnabled: false
-    }) as ElementField;
-
-    const licenseRatingField = this.createElementField({
-      Element: lisanceElement,
-      Name: "License Rating",
+    // Child element - Rating field
+    const childRatingField = this.createElementField({
+      Element: childElement,
+      Name: "Rating",
       DataType: ElementFieldDataType.Decimal,
       UseFixedValue: false,
       RatingEnabled: true,
       SortOrder: 1
     }) as ElementField;
 
-    // Restricted License Item
-    const restrictedLicense = this.createElementItem({
-      Element: lisanceElement,
-      Name: "Restricted License"
+    // Child element - Item 1
+    const childItem1 = this.createElementItem({
+      Element: childElement,
+      Name: "Child item 1"
     }) as ElementItem;
 
-    // Cells
+    // Child element - Cell 1
     this.createElementCell({
-      ElementField: rightToUseField,
-      ElementItem: restrictedLicense,
-      StringValue: "Yes"
+      ElementField: childRatingField,
+      ElementItem: childItem1,
     });
 
-    this.createElementCell({
-      ElementField: rightToCopyField,
-      ElementItem: restrictedLicense,
-      StringValue: "No"
-    });
-
-    this.createElementCell({
-      ElementField: rightToModifyField,
-      ElementItem: restrictedLicense,
-      StringValue: "No"
-    });
-
-    this.createElementCell({
-      ElementField: rightToSellField,
-      ElementItem: restrictedLicense,
-      StringValue: "No"
-    });
-
-    this.createElementCell({
-      ElementField: licenseRatingField,
-      ElementItem: restrictedLicense,
-    });
-
-    // Open Source License Item
-    const openSourceLicense = this.createElementItem({
-      Element: lisanceElement,
-      Name: "Open Source License"
+    // Child element - Item 2
+    const childItem2 = this.createElementItem({
+      Element: childElement,
+      Name: "Child item 2"
     }) as ElementItem;
 
-    // Cells
+    // Child element - Cell 2
     this.createElementCell({
-      ElementField: rightToUseField,
-      ElementItem: openSourceLicense,
-      StringValue: "Yes"
+      ElementField: childRatingField,
+      ElementItem: childItem2,
     });
 
-    this.createElementCell({
-      ElementField: rightToCopyField,
-      ElementItem: openSourceLicense,
-      StringValue: "Yes"
-    });
+    // Parent element
+    const parentElement = project.ElementSet[0];
 
-    this.createElementCell({
-      ElementField: rightToModifyField,
-      ElementItem: openSourceLicense,
-      StringValue: "Yes"
-    });
-
-    this.createElementCell({
-      ElementField: rightToSellField,
-      ElementItem: openSourceLicense,
-      StringValue: "Yes"
-    });
-
-    this.createElementCell({
-      ElementField: licenseRatingField,
-      ElementItem: openSourceLicense,
-    });
-
-    // Main element
-    const mainElement = project.ElementSet[0]
-
-    const licenseField = this.createElementField({
-      Element: mainElement,
-      Name: "License",
+    // Parent element - Child field
+    const childField = this.createElementField({
+      Element: parentElement,
+      Name: "Child",
       DataType: ElementFieldDataType.Element,
     }) as ElementField;
+    childField.SelectedElement = childElement;
 
-    licenseField.SelectedElement = lisanceElement;
+    // Parent element - Item 1
+    parentElement.ElementItemSet[0].Name = "Parent item 1";
 
-    mainElement.ElementItemSet[0].Name = "Hidden Knowledge"
-
+    // Parent element - Cell 1
     this.createElementCell({
-      ElementField: licenseField,
-      ElementItem: mainElement.ElementItemSet[0]
+      ElementField: childField,
+      ElementItem: parentElement.ElementItemSet[0]
     });
-    mainElement.ElementItemSet[0].ElementCellSet[0].SelectedElementItem = restrictedLicense;
+    parentElement.ElementItemSet[0].ElementCellSet[0].SelectedElementItem = childItem1;
 
-    mainElement.ElementItemSet[1].Name = "True Source"
+    parentElement.ElementItemSet[1].Name = "Parent item 2";
     this.createElementCell({
-      ElementField: licenseField,
-      ElementItem: mainElement.ElementItemSet[1]
+      ElementField: childField,
+      ElementItem: parentElement.ElementItemSet[1]
     });
-    mainElement.ElementItemSet[1].ElementCellSet[0].SelectedElementItem = openSourceLicense;
+    parentElement.ElementItemSet[1].ElementCellSet[0].SelectedElementItem = childItem2;
 
     // Return
     return project;
@@ -606,6 +453,35 @@ export class ProjectService {
 
   updateElementCellDecimalValue(elementCell: ElementCell, value: number) {
     // Todo Implement!
+  }
+
+  // Creates a project with one element and x number of elements
+  private createDefaultProject(projectName: string, mainElementName: string, numberOfItems: number): Project {
+
+    // Project
+    const project = this.appEntityManager.createEntity("Project", {
+      User: this.authService.currentUser,
+      Name: projectName,
+      Origin: "http://"
+    }) as Project;
+
+    // Main element
+    const element = this.createElement({
+      Project: project,
+      Name: mainElementName
+    }) as Element;
+
+    // Items, cells
+    for (let i = 0; i < numberOfItems; i++) {
+
+      const item = this.createElementItem({
+        Element: element,
+        Name: `Item ${i + 1}`
+      }) as ElementItem;
+    }
+
+    // Return
+    return project;
   }
 
   private getUpdateComputedFieldsUrl(projectId: number) {