+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ### Changelog
 
+**0.7.4**
+
+* AngularClient - Create project, user element cell and field default values fix
+
 **0.7.3**
 
-* AngularClient - Two Elements example
+* AngularClient - Create Project - "Parent - Child" template
 
 **0.7.2**